Use better algorithm to resolve presence; rename resource to session.
Contact.resolvePresence used to use a switch statement to find the
highest priority presence to use (favoring, for instance, available over
xa). Now we'll use sort() to sort the sessions by the index of their
presence in the ordered presence priority array and access the highest
priority presence wit the 0 index. Much less code and less room for error.