Use multiple promises to consolidate approval and reception of offer.
This makes it easy for us to combine two code paths into one. We
have to make sure the user has approved the call (or the developer has
chosen to skip approval) before we call setRemoteDescription.  This way,
we can ensure that we don't process the offer until both of these
conditions have been met, and we don't have to keep checking state
variables in if statements to manage it.