Accept callbacks to Call constructor and Call.answer.
Add onLocalVideo, onRemoteVideo, and onHangup callbacks. Accept in
constructor for the initiator and in Call.answer for the non-initiator.
This should eliminate the possibility of race conditions where the
developer listens to these three events after they are fired, since none
of these events is fired until after answer() is called.