Hang up automatically and warn if no call or connection handlers.
Previously a call or direct connection would just hang if the developer
forgot to register handlers to accept the connections. Now on the
receiving side, we can detect if the developer has added handlers to
handle the event. If they haven't, we reject the call or direct
connection and print a warn statement to the console.