Implement reconnect without socket.io, added manual heartbeat.
Because the token has to be on the socket.io socket URL, we cannot use
socket.io's built-in reconnect. Thus, we have to disable it and listen
to disconnect so we can manually create another socket and connect
again. Also, we don't get notified when the token is disabled, and it
may become invalidated for reasons other than the TTL, so we must
implement a manual heartbeat to detect  when we are disconnected and
reconnect again. This commit also required some reworking of event
listeners to better prevent double listener attachment.