Use done() carefully:
Use done() & don't return the promise if the dev provides handlers
Use done() if we would never return the promise.
Don't use done() and do return the promise if the dev doesn't provide
handlers.
Don't use done() and do return the promise if it's an internal method
that doesn't take handlers.