Fix some bugs:
Fix bug where lib adding onSuccess masks whether dev is adding one, thus
promise is never returned.
Improvement in which we check to see if we are connected so we don't
throw unrelated errors when we are not connected.
Fix bug where we were rejecting a promise instead of a deferred.