remove object prototype additions and add combine task for easy debugging
the object prototype property ‘clone’ was clashing with underscore’s ‘clone’ method, causing an exception. I removed the defined property as well as the isNumber property, because it was not being used by anything except for outdated tests. I also added a ‘grunt combine’ task so that I could test the brightstream library without having to rely on a source map, which, unless the contents of the source map are embedded in the minified file, isn’t very useful without each individual file served in the directory along with the concat+minified version of the files.