Fix a reconnect bug related to timeouts & opening more than 1 socket.
The way we were kicking off a manual reconnect on error, disconnect, and
connect_failed makes it possible for transporter to get into a state
where it leaks websockets. Now we have an isConnecting flag so we never
try to open a websocket while a connection is open or in progress. Also,
the manual reconnect timeout started at 500ms while the socket.io
connect timeout was 2000ms, so it's possible for transporter to try
again while we still don't know if the last attempt was successful or
not. Also removed some socket.io configuration options that no longer
apply because we are not utilizing socket.io's built-in reconnect.