start saucer-section for functional tests
To be able to run the functional tests in transporter saucer-section
needs to be running as collective talks to it for terms of service.

the following grunt tasks were added:
  start-saucer-section, stop-saucer-section
the following config options were added:
  saucerSection.dir and saucerSection.port
they default to '../../../saucer-section' and 3000

grunt.util.spawn is used to start saucer-section with the grunt default
task in saucer-section

A jira task has been created to enable tranporter functional tests to be
run in the full environment:  MER-1355.