Send App-Token for all HTTP(S) requests
During disconnects, I noticed a lot of 401's for DELETE
/v1/appauthsessions. This was because even though we had an App-Token,
it wasn't being put into the request headers.

There was some cumbersome exception eating code I also got rid of, since
those failure paths are basically impossible for us to hit (we just
opened the XHR, and we know the values we're adding to the XHR are
valid). If there is an exception, it's better to bubble it up instead of
eating it.