WebSocket request tracking
This patch removes the existing heartbeat and timeout code for dealing
with WebSocket errors, and replaces it with request tracking. The
signaling channel keeps track of every pending request on the WebSocket,
and will fail any pending requests if the WebSocket disconnects.

Since the danger is now that we'll have hanging requests that never time
out, the request logging was improved to include response times. XHR
logging messages were changed to be consistent.

A few minor cleanups are along for the ride.