Remove the checking for 'system' channels.
This is a remnant from how messages originating from the routing service
used to be delivered. Since 'channel' is a user controlled field, if
they create a group which happens to have the word 'system' in the name,
it will be handled as not being associated with a channel.

This is something we happened to notice with Apollo.