Fix EventEmitter bug with `once`.
The EventEmitter was using Function.toString() to compare functions.
When the listener was added with once(), the function's toString output
would always be the same string, so subsequent calls to once() would
cause the listener not to be added. This patch sets a toString()
property on the listener with the value of the original listener if it
was added with once().