don't attempt to setup temporary RemoteMedia
In Google Chrome, attempting to set the 'src' property of a video
element to `undefined` or `null` will cause the browser to issue
a request to __page_root__/null, i.e. 'http://localhost:9876/null'.
This patch guards against this by not attempting to call
attachMediaStream when we know that the RemoteMedia it is setting
up is temporary, and will therefore have no stream.