Pin test library versions to address explosions
This patch pins grunt-karma at v0.11.x and karma-webpack at v1.6.x to
address karma-webpack v1.7.x incompatibility with karma < 0.13.x. Could
not upgrade karma to 0.13.x at this time because it is incompatible with
the use of urlRoot in the configuration file, which we use to resolve
some of our test dependencies like karma-chai-sinon and karma-dirty-chai.