Optimize group leave handling
When we receive a group leave event, we need to do some garbage
collecting, in case the endpoint is no longer referenced by any other
groups.

This was handled by requesting updated group lists on all groups we know
about. Unfortunately, while very reliable, this caused a lot of extra
requests and work re-processing the group lists.

This patch adds a bit of bookkeeping on the endpoint to determine
whether it's in any groups or not.