Remove client#error event documentation
This patch removes the documentation referencing client#error event.
This event was probably removed at some point, but our documentation
still referenced it as the canonical way to listen for a connection
failure. The docs have been updated to reflect that passing onSuccess
and onError params to the connect function is the canonical way, but
you can also add a failure handler to the returned promise instead.

Related to MER-4387.