Fix client docs upload to s3
The s3 uploader script just uploads the files in the same structure
they are on disk. The main documentation site expects the client
docs to be in a directory called 'js-library', so we need to add that
to the path created by the documentation generator so the uploader
can put the files in the right place on s3.