Simply test suite and remove functional tests
Functional tests are run in a separate test suite in our bamboo
project, so we're removing them from this project for now because
we cannot afford to maintain them in two places and the ones in this
project had been neglected for a long time anyway. On top of that,
the test suite had a separate build process for the Respoke.js files,
which kind of flies in the face of actually testing what we are
deploying. This new setup tests the same built files that are deployed.