Generate unminified distributable, fix source-maps
Previously the source maps were very difficult to use because the
underlying uglified source was combining multiple lines due to the
'compress' option. This made breakpoints very difficult to set,
because lines in the sourcemap were not mapping 1:1 with the lines
in the underlying source. This patch removes the compression option.

This patch also generates a new `respoke.js` build file that is the
built respoke library without the uglification.

Finally, many unneeded dev dependencies have been removed.