Attempt a sdp munge on setRemoteDescription fail
When Chrome 46 was released, they started using UDP/TLS/RTP/SAVPF as
the default transport protocol on the m= line of the sdp generated by
createOffer. This broke inter-op with nw.js 0.12.2, which uses Chromium
41.

This patch attempts to alter the SDP when setRemoteDescription fails to
change the transport protocol from UDP/TLS/RTP/SAVPF to simply
RTP/SAVPF, which was what was sent previously, and re-try the
setRemoteDescription call. If the sdp does not contain this protocol,
the setRemoteDescription is not re-tried. If the setRemoteDescription
fails on the second try, the same old error handling logic will be
triggered.

Related to MER-4440.