Prevent attachMediaStream if stream doesn't exist
When you try to attach an `undefined` or `null` mediaStream to a
video element, it makes a request to __host__/null, i.e.
http://localhost/null to try to fetch the resource. We believe that this
is causing our tests to sporadically fail, and it's useless anyway, so
this patch makes sure that doesn't happen.

See https://code.google.com/p/chromium/issues/detail?id=378433