Improve Call.sendTones docs, Ninja refactor
Previously the sendTones docs displayed on docs.respoke.io didn't
mention DTMF at all, which makes it difficult to understand what
sendTones is meant to be used for. This patch fixes this, as well as
cleans up the struture of the sendTones methods in peerConnection and
some misc other docs fixes.

Related to MER-4458.