Allow metadata to be passed in offer signals
This patch is a step towards allowing metadata to be passed to
chan_respoke endpoints, as they have a particularly difficult time
bridging the gap between the web context and the phone system. With
these changes, a user can now pass a `metadata` property to any of the
methods to start a call, and that metadata will be present on the call
or directConnection object on the receiver side. The metadata property
is only placed on the 'offer' signal.

Mixed in this commit is also a bunch of trivial cleanup of unused
properties and other misc tiny things.

Related to MER-4465