Fix renegotiation for same browser aud/vid calling
This patch builds on the previous work done to implement renegotiation
in Respoke to allow adding and removing video and audio to/from a call.

Firefox
--------

With this patch, two Firefox 43 browsers can renegotiation to add or
remove audio or video from either side without issue.

Chrome
--------

With this patch, two Chrome 49 browsers can renegotiation to add or
remove audio or video from either side without issue. The only caveat is
when the callee tries to renegotiate a video-only call to an audio+video
call, which causes Chrome to throw an exception due to a DTLS role swap.
The corresponding chrome bug tracker issue is located at
https://bugs.chromium.org/p/webrtc/issues/detail?id=2782 . The only
known workaround for this bug is for the caller to munge it's answer sdp
to set the correct dtls role prior to sending to the callee.

Chrome versions prior to 49 have a bug when removing certain media that
causes the renegotiator to set the direction of the media that it is
removing as `a=sendonly` when it should be `a=recvonly`. This bug causes
the renegotiatee to stop sending its media. The only known workaround
for this is to intelligently set the options passed to `createOffer` to
cause the offer to have the correct direction flag applied to it.

Firefox<->Chrome Interop
-------------------------

There are currently numerous issues with cross browser interop with this
patch that are as yet unresolved. Known problems are:

 - when Chrome initiates a renegotiation on a call that Firefox had
   started, Firefox throws a InvalidSessionDescriptionError indicating
   that it does not support ice restart. Affects at least all Chrome
   versions up to 49, and passing { iceRestart: false } as an offer
   option on the Chrome side does not fix it.

 - when Firefox initiates a renegotiation on a call that Chrome had
   started, adding video to the call, the video track on the Chrome side
   is muted with no known way to unmute it.

 - when video is removed by Firefox, the video in Chrome gets stuck on
   the last frame sent by Firefox.

 - A few other smaller issues, such as the state machine not properly
   resetting connection timers, invalid state transitions, and call
   debugs not being sent due to exceeding the maximum body size for a
   post to the Respoke API.

Breaking API Changes
---------------------

 - `LocalMedia` (outgoing) streams are no longer automatically reused
   between calls. If this functionality is desired, the `outgoingMedia`
   property should be set when starting a new call, passing in a started
   `LocalMedia` instance. Additionally, `respoke.streams` (where the
   existing streams were previously "cached") no longer exists. This
   caching was removed because, in it's existing form, it could not
   differentiate between two streams obtained using the same
   constraints. This caused problems when trying to change from one
   video source to another.

 - Revoking permission to a stream, thereby stopping the stream, no
   longer automatically hangs up the call if it was the only stream left
   on the call. This functionality was interfering with renegotiation
   because prior to re-offering on Chrome, you must stop and remove a
   stream prior to adding the new stream. This was most noticable when
   performing a screen share and then pressing the "Stop Sharing" button
   that Chrome shows while sharing your screen. For anyone that was
   relying on that functionality, it can still be emulated by listening
   to the 'stop' event on their `LocalMedia` instance. In the handler,
   a check can be done against the call's `outgoingMediaStreams` and
   `incomingMediaStreams` to see if they are empty, and if they are,
   hangup the call.

API Additions
--------------

 - A 'remote-media-received' event will now be fired on the `Call`
   object any time new remote media is added to a call. This event will
   overlap with the 'connect' event upon receiving the first remote
   media, but only 'remote-media-received' will be fired upon receiving
   new media due to renegotiation.

 - A new param, `onRemoteMedia`, can now be passed when starting a call,
   to register a 'remote-media-received' event handler.

 - A few private methods have been added to a `Call` to support
   renegotiating media:

    - `call.addVideo()` allows you to add video,
    - `call.removeVideo()` allows you to remove video,
    - `call.addAudio()` allows you to add audio, and
    - `call.removeAudio()` allows you to remove audio.

   If the call's media is not in the correct state for the method when
   called, it will automatically resolve the returned promise. These
   methods are private for now until renegotiation is released as a
   stable feature.

The work here is a solid "second pass", but by no means a bulletproof
implementation of renegotiation. Screen sharing remains unsupported by
this implementation. Additionally, unit tests are lacking for the new
features added. Regardless, it should be a solid foundation to work from
when moving forward to implementing features such as mobile camera
swapping and video conferencing.

Paired with Jeff Parrish.